refactor: Grouping each feature and Deleted Tests Dir, Split PlanListView
