+++ LetsEat/LetsEat/LaunchView/LoginView.swift
@@ -1,0 +1,20 @@
+//
+//  LoginView.swift
+//  LetsEat
+//
+//  Created by 이현재 on 2023/09/11.
+//
+
+import SwiftUI
+
+struct LoginView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct LoginView_Previews: PreviewProvider {
+    static var previews: some View {
+        LoginView()
+    }
+}
