+++ LetsEat/LetsEat/Manager/AppState.swift
