+++ LetsEat/LetsEat/Model/Plan.swift
@@ -1,0 +1,14 @@
+//
+//  Plan.swift
+//  LetsEat
+//
+//  Created by 이현재 on 2023/09/11.
+//
+
+import Foundation
+
+struct Plan: Hashable {
+    var name: String
+    var startDate: Date
+    var endDate: Date
+}
