--- LetsEat/LetsEatUITests/LetsEatUITestsLaunchTests.swift
@@ -1,32 +1,0 @@
-//
-//  LetsEatUITestsLaunchTests.swift
-//  LetsEatUITests
-//
-//  Created by 이현재 on 2023/08/28.
-//
-
-import XCTest
-
-final class LetsEatUITestsLaunchTests: XCTestCase {
-
-    override class var runsForEachTargetApplicationUIConfiguration: Bool {
-        true
-    }
-
-    override func setUpWithError() throws {
-        continueAfterFailure = false
-    }
-
-    func testLaunch() throws {
-        let app = XCUIApplication()
-        app.launch()
-
-        // Insert steps here to perform after app launch but before taking a screenshot,
-        // such as logging into a test account or navigating somewhere in the app
-
-        let attachment = XCTAttachment(screenshot: app.screenshot())
-        attachment.name = "Launch Screen"
-        attachment.lifetime = .keepAlways
-        add(attachment)
-    }
-}
